organizing MarkdownInterpreter folder
