+++ src/components/MarkdownInterpreter/MarkdownElement/MarkdownElementNewline.vue
@@ -1,0 +1,14 @@
+<template>
+	<br :v-repeat="element.content[0]">
+</template>
+
+<script>
+	export default {
+		props: {
+			element: {
+				type: Object,
+				required: true
+			}
+		}
+	}
+</script>
