+++ src/components/MarkdownInterpreter/MarkdownElement/MarkdownElementText.vue
@@ -1,0 +1,14 @@
+<template>
+	<span>{{ element.content }}</span>
+</template>
+
+<script>
+	export default {
+		props: {
+			element: {
+				type: Object,
+				required: true
+			}
+		}
+	}
+</script>
