--- src/components/MarkdownInterpreter/MarkdownElementNewline.vue
@@ -1,14 +1,0 @@
-<template>
-	<br :v-repeat="element.content[0]">
-</template>
-
-<script>
-	export default {
-		props: {
-			element: {
-				type: Object,
-				required: true
-			}
-		}
-	}
-</script>
