--- src/components/MarkdownInterpreter/MarkdownElementText.vue
@@ -1,14 +1,0 @@
-<template>
-	<span>{{ element.content }}</span>
-</template>
-
-<script>
-	export default {
-		props: {
-			element: {
-				type: Object,
-				required: true
-			}
-		}
-	}
-</script>
