+++ src/components/MarkdownInterpreter/MarkdownParser/MarkdownTokenScannerImage.js
